backend/requirements.txt
fastapi
uvicorn[standard]
boto3
pyyaml
pydantic-settings